requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
amqp==5.3.1
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
billiard==4.2.4
black==26.3.1
boto3==1.42.92
botocore==1.42.92
celery==5.6.3
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==6.0.4
django-anymail==15.0
django-cors-headers==4.9.0
django-debug-toolbar==6.3.0
django-environ==0.13.0
django-filter==25.2
django-ratelimit==4.1.0
django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
elastic-transport==8.17.1
elasticsearch==8.19.3
elasticsearch-dsl==8.18.0
flake8==7.3.0
frozenlist==1.8.0
gunicorn==25.3.0
hiredis==3.3.1
idna==3.12
inflection==0.5.1
iniconfig==2.3.0
jmespath==1.1.0
kombu==5.6.2
mccabe==0.7.0
multidict==6.7.1
mypy_extensions==1.1.0
packaging==26.1
pathspec==1.0.4
pillow==12.2.0
platformdirs==4.9.6
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.12
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.3
pytest-django==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytokens==0.4.1
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.4.0
requests==2.33.1
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.12.0
yarl==1.23.0
